refactor: extract terminal-bench adapter cli commands
